pyproject.toml
[project]
name = "codemind"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.6.3",
    "fastapi==0.116.1",
    "httpx==0.27.0",
    "jinja2>=3.1.0",
    "litellm>=1.30.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "python-dotenv==1.1.1",
    "pyyaml>=6.0",
    "redis>=7.4.0",
    "tree-sitter>=0.25.2",
    "tree-sitter-go>=0.25.0",
    "tree-sitter-java>=0.23.5",
    "tree-sitter-javascript>=0.25.0",
    "tree-sitter-python>=0.25.0",
    "tree-sitter-typescript>=0.23.2",
    "uvicorn==0.35.0",
]
